--- sdk/basyx/adapter/__init__.py
@@ -1,9 +1,0 @@
-"""
-This package contains different kinds of adapters.
-
-* :ref:`json <adapter.json.__init__>`: This package offers an adapter for serialization and deserialization of BaSyx
-  Python SDK objects to/from JSON.
-* :ref:`xml <adapter.xml.__init__>`: This package offers an adapter for serialization and deserialization of BaSyx
-  Python SDK objects to/from XML.
-* :ref:`aasx <adapter.aasx>`: This package offers functions for reading and writing AASX-files.
-"""
